+++ src/pages/home/calendar/confirmedTimes.js
@@ -1,3 +1,4 @@
+// Placeholder for admin-created events before admin function integration
 const confirmedTimes = [
   new Date(2025, 1, 13, 14, 0),
   new Date(2025, 1, 13, 14, 30),
